split http evidence into multiple files
